Created Footer component to hold footer that was in the App.
