+++ src/components/ControlsContent/ControlsContent.tsx
@@ -8,25 +8,11 @@ import Notes from "../Notes";
 export default function ControlsContent(): JSX.Element {
   const classes = useStyles();
   const [showSettings, setShowSettings] = React.useState(false);
-  // const { notes, setNotes } = useData()!; // eslint-disable-line
-
-  // const debounce = (func: Function, wait: number) => {
-  //   let timeout: ReturnType<typeof setTimeout>;
-
-  //   return function executedFunction(...args) {
-  //     const later = () => {
-  //       clearTimeout(timeout);
-  //       func(...args);
-  //     };
-
-  //     clearTimeout(timeout);
-  //     timeout = setTimeout(later, wait);
-  //   };
-  // };
 
   return (
     <div id="controls" className={classes.root}>
       <ScoresPanel />
+
       <Notes />
 
       <Button
