+++ src/components/Footer/Footer.styles.ts
@@ -1,0 +1,8 @@
+import { ITheme } from "utils/types";
+import { createUseStyles } from "react-jss";
+
+export default createUseStyles((theme: ITheme) => ({
+  root: {
+    backgroundColor: theme.backgroundColor,
+  },
+}));
