+++ src/components/Notes/Notes.styles.ts
@@ -2,14 +2,17 @@ import { ITheme } from "utils/types";
 import { createUseStyles } from "react-jss";
 
 export default createUseStyles((theme: ITheme) => ({
-  root: {},
-  notes: {
+  root: (props) => ({
+    margin: props.isMobile ? "0.5rem 1rem" : "",
+  }),
+  notes: (props) => ({
     fontSize: "1.25rem",
     width: "100%",
-    minHeight: "350px",
+    minHeight: props.isMobile ? "60vh" : "350px",
+    maxHeight: props.isMobile ? "60vh" : "",
     height: "100%",
     backgroundColor: theme.neutralBackgroundColor,
-  },
+  }),
   notesContainer: {
     flexGrow: 1,
   },
